services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.5.2
sqlalchemy==2.0.34
psycopg[binary]==3.2.3
python-dotenv==1.0.1
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0

# tests
pytest==8.3.3
httpx==0.27.2


